get rid of UserGroup controller
