+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/RouteConfig.kt
@@ -4,7 +4,7 @@ import org.vaccineimpact.orderlyweb.*
 import org.vaccineimpact.orderlyweb.app_start.routing.api.*
 import org.vaccineimpact.orderlyweb.app_start.routing.web.*
 import org.vaccineimpact.orderlyweb.controllers.web.IndexController
-import org.vaccineimpact.orderlyweb.controllers.web.UserGroupController
+import org.vaccineimpact.orderlyweb.controllers.web.AdminController
 
 interface RouteConfig
 {
@@ -26,7 +26,7 @@ object WebRouteConfig : RouteConfig
     private val metricsEndpoint = WebEndpoint("/metrics/", IndexController::class, "metrics")
             .json()
 
-    private val adminEndpoint = WebEndpoint("/admin/", UserGroupController:: class, "admin")
+    private val adminEndpoint = WebEndpoint("/admin/", AdminController:: class, "admin")
             .secure(setOf("*/users.manage"))
 
     override val endpoints: List<EndpointDefinition> =
