+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/routing/web/WebUserRouteConfig.kt
@@ -3,7 +3,7 @@ package org.vaccineimpact.orderlyweb.app_start.routing.web
 import org.vaccineimpact.orderlyweb.*
 import org.vaccineimpact.orderlyweb.app_start.RouteConfig
 import org.vaccineimpact.orderlyweb.controllers.web.UserController
-import org.vaccineimpact.orderlyweb.controllers.web.UserGroupController
+import org.vaccineimpact.orderlyweb.controllers.web.AdminController
 import spark.route.HttpMethod
 
 object WebUserRouteConfig : RouteConfig
@@ -26,27 +26,9 @@ object WebUserRouteConfig : RouteConfig
                     .json()
                     .transform()
                     .secure(usersManage),
-            WebEndpoint("/user-groups/:user-group-id/actions/associate-permission/",
-                    UserGroupController::class, "associatePermission",
+            WebEndpoint("/users/:user-id/actions/associate-permission/",
+                    AdminController::class, "associatePermission",
                     method = HttpMethod.post)
-                    .json()
-                    .transform()
-                    .secure(usersManage),
-            WebEndpoint("/user-groups/",
-                    UserGroupController::class, "addUserGroup",
-                    method = HttpMethod.post)
-                    .json()
-                    .transform()
-                    .secure(usersManage),
-            WebEndpoint("/user-groups/:user-group-id/",
-                    UserGroupController::class, "addUser",
-                    method = HttpMethod.post)
-                    .json()
-                    .transform()
-                    .secure(usersManage),
-            WebEndpoint("/user-groups/:user-group-id/user/:email",
-                    UserGroupController::class, "removeUser",
-                    method = HttpMethod.delete)
                     .json()
                     .transform()
                     .secure(usersManage),
