+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/integration_tests/tests/api/UserTests.kt
@@ -1,3 +1,4 @@
+
 package org.vaccineimpact.orderlyweb.tests.integration_tests.tests.api
 
 import org.assertj.core.api.Assertions
