+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/unit_tests/controllers/web/AdminControllerTests.kt
@@ -1,0 +1,30 @@
+package org.vaccineimpact.orderlyweb.tests.unit_tests.controllers.web
+
+import com.nhaarman.mockito_kotlin.*
+import org.assertj.core.api.Assertions
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.Test
+import org.mockito.ArgumentCaptor
+import org.mockito.internal.verification.Times
+import org.vaccineimpact.orderlyweb.ActionContext
+import org.vaccineimpact.orderlyweb.controllers.web.AdminController
+import org.vaccineimpact.orderlyweb.controllers.web.RoleController
+import org.vaccineimpact.orderlyweb.db.AuthorizationRepository
+import org.vaccineimpact.orderlyweb.errors.MissingParameterError
+import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
+import org.vaccineimpact.orderlyweb.test_helpers.TeamcityTests
+import org.vaccineimpact.orderlyweb.viewmodels.Breadcrumb
+import org.vaccineimpact.orderlyweb.viewmodels.IndexViewModel
+
+class AdminControllerTests : TeamcityTests()
+{
+    @Test
+    fun `returns correct breadcrumbs for admin page`()
+    {
+        val sut = AdminController(mock())
+        val model = sut.admin()
+        assertThat(model.breadcrumbs).containsExactly(IndexViewModel.breadcrumb,
+                Breadcrumb("Admin", "http://localhost:8888/admin"))
+    }
+
+}
