+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/unit_tests/controllers/web/UserControllerTests.kt
@@ -1,15 +1,18 @@
 package org.vaccineimpact.orderlyweb.tests.unit_tests.controllers.web
 
-import com.nhaarman.mockito_kotlin.doReturn
-import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockito_kotlin.*
+import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import org.mockito.ArgumentCaptor
 import org.vaccineimpact.orderlyweb.ActionContext
+import org.vaccineimpact.orderlyweb.controllers.web.RoleController
 import org.vaccineimpact.orderlyweb.controllers.web.UserController
 import org.vaccineimpact.orderlyweb.db.AuthorizationRepository
 import org.vaccineimpact.orderlyweb.db.UserRepository
 import org.vaccineimpact.orderlyweb.models.User
 import org.vaccineimpact.orderlyweb.models.permissions.PermissionSet
+import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
 import org.vaccineimpact.orderlyweb.test_helpers.TeamcityTests
 
 class UserControllerTests : TeamcityTests()
@@ -162,4 +165,58 @@ class UserControllerTests : TeamcityTests()
         assertThat(result[3].displayName).isEqualTo("r4@email.com")
     }
 
+
+    @Test
+    fun `adds permission to user`()
+    {
+        val actionContext = mock<ActionContext> {
+            on { this.params(":user-id") } doReturn "user1@example.com"
+            on { this.postData() } doReturn mapOf(
+                    "action" to "add",
+                    "name" to "test.permission",
+                    "scope_prefix" to "report",
+                    "scope_id" to "report1"
+            )
+        }
+
+        val authRepo = mock<AuthorizationRepository>()
+        val sut = UserController(actionContext, mock(), authRepo)
+        val result = sut.associatePermission()
+
+        assertThat(result).isEqualTo("OK")
+
+        val permissionCaptor: ArgumentCaptor<ReifiedPermission> = ArgumentCaptor.forClass(ReifiedPermission::class.java)
+        verify(authRepo).ensureUserGroupHasPermission(eq("user1@example.com"), capture(permissionCaptor))
+
+        val permission = permissionCaptor.value
+        Assertions.assertThat(permission.name).isEqualTo("test.permission")
+        Assertions.assertThat(permission.scope.value).isEqualTo("report:report1")
+    }
+
+    @Test
+    fun `removes permission from user`()
+    {
+        val actionContext = mock<ActionContext> {
+            on { this.params(":user-id") } doReturn "user1@example.com"
+            on { this.postData() } doReturn mapOf(
+                    "action" to "remove",
+                    "name" to "test.permission",
+                    "scope_prefix" to "report",
+                    "scope_id" to "report1"
+            )
+        }
+
+        val authRepo = mock<AuthorizationRepository>()
+        val sut = UserController(actionContext, mock(), authRepo)
+        val result = sut.associatePermission()
+
+        assertThat(result).isEqualTo("OK")
+
+        val permissionCaptor: ArgumentCaptor<ReifiedPermission> = ArgumentCaptor.forClass(ReifiedPermission::class.java)
+        verify(authRepo).ensureUserGroupDoesNotHavePermission(eq("user1@example.com"), capture(permissionCaptor))
+
+        val permission = permissionCaptor.value
+        Assertions.assertThat(permission.name).isEqualTo("test.permission")
+        Assertions.assertThat(permission.scope.value).isEqualTo("report:report1")
+    }
 }
